Consolidate typeface stuff into TypefaceCompat.kt
